src/root_layer/root_fl_manager/requirements.txt
flask ~= 3.0.2
flask-openapi3 ~= 3.0.2
requests ~= 2.28.1
docker ~= 7.0.0
gitpython ~= 3.1.42
paho-mqtt == 2.0
PyGithub == 2.2.0
pymongo ~= 4.6.3
icecream
